Refactored file architecture for implementation of simulation
